swap import and export icons (#212)
These icons are designed from the perspective of the grid.

"tower-import" means the grid is importing power. From the SigenStor perspective, this is when the SigenStor is sending/exporting power to the grid.

So when the local system is 'exporting', we need to use the 'tower-import' icon. And vice versa for importing.